requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta-lib>=0.4.26
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
reportlab>=4.0.0
pyyaml>=6.0
